FINALE

+++ database/migrations/2023_06_02_160707_create_feed_backs_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('id_visit')->constrained('visits');
             $table->foreignId('id_question')->constrained('questions');
-            $table->unsignedTinyInteger('skala_feed'); // Menggunakan tipe data unsignedTinyInteger
+            $table->float('skala_feed'); // Menggunakan tipe data unsignedTinyInteger
             $table->timestamps();
         });
     }
